Modify untilt to update with new Z.
While testing, I noticed that issuing M653, M654, M653 in sequence
would E side axis on the second M653, not the Z axis. I then did an
M654 and tried to move the platform up by 1mm, and instead it went
down 19mm. Because the Z position had never been updated, it thought
it was lower than it was, and moves back to the destination. This is
potentially bad, because someone could say "go up 10mm" to make it
easier to get their part out, but it thinks it should be 20mm below
where it actually is, so it goes down 10mm and makes it worse. breaks
things. Let's not do that, and instead save away our current position
when we go back to level.

Now, if someone hand cranks us to level, it's a crapshoot, because
they could either crank Z down, or E up, and we have no way of
knowing. But, then they'd likely just print anyway, and the M651
command will clear the tilted state and rehome anyway, so we should be
okay.