Make MarlinSerial able to handle multiple serial ports.
The only thing which needs to be port-specific at this point is ISR
registration, which calls .isr() in the correct object based on the
signal received. This seems to be limited by avr-libc, which uses
predefined names to do macro magic, and those names need to be
known at compile time.