Replace M651 with M699.
We've decided to mainline M651, since there is a backwards compatible
option. Namely, if you don't set H, then it will be set to 0, which
means it will return back to the original Z position, and then you can
still use your existing G1 command to go up that layer.