Update Configuration.h
Disable MAX Endstops for workaround to what appears to be a Marlin bug. When using a G92 to set the Z axis coordinates the system refuses to move in the Z+ direction until homing is processed and the endstop depressed. This isn't going to work for compatibility with printing resume functions so it has to be changed until the true issue can be resolved, if possible.